requirements.txt
python-splitwise
fastapi
uvicorn